Resubmission: Minor revisions
Based on the advice received, the Editor feels that your manuscript
could be accepted for publication should you be prepared to incorporate
minor revisions.  When preparing your revised manuscript, you are asked
to carefully consider the reviewer comments which are attached, and
submit a list of responses to the comments.  Your list of responses
should be uploaded as a file in addition to your revised manuscript.

COMMENTS FOR THE AUTHOR:

The Associate Editor agrees that the paper is in a much better shape.
The points raised by the referees are reasonably minor. I therefore
recommend the paper to be published subject to minor corrections.

Reviewer #2: 1) In the abstract sentence "This paper will examine the
latter approach and present a systematic approach for doing so
effectively" is not clear. Please clarify the contributions of this
paper.

2) Section 9 is better to be bisected to sections of "Discussion" and
"Conclusions".

Reviewer #3: The authors have revised the paper accordingly, apart from
lack of future works.  It would be interesting to see the proposed work
applied to service-oriented software engineering or cloud services to
schedule the services in order to meet the QoS. The authors can use the
following reference to discuss.

Cheikhrouhou, S., Kallel, S., Guermouche, N. et al. SOCA (2015), "The
temporal perspective in business process modeling: a survey and research
challenges",  9: 75. doi:10.1007/s11761-014-0170-x